Store camera position for later state restore
State is restored on undo, restart, or when loading a level
after already hitting a checkpoint in that level. We now store
the target camera position after every move so that when we
do any of those restores, we can go back to the same camera
position. That means undo now puts the camera back into the
position it was in when you were in that state. Similarly,
checkpoints will remember the camera position.

Figuring out how to do this nicely also resulted in a neater
implementation of hooking into the level load process.

Signed-off-by: Joseph Mansfield <sftrabbit@gmail.com>