Improved level select. Fixing known bugs.
If level select is enabled, don't delete save data when selecting "New game". Level select cursor starts on the last played level (instead of the earliest level that is yet to be completed). Other tweaks to prevent level select scrolling from firing you out of bounds and to non-exising levels.