#31 Fixed capitalization for some of the error messages
As part of the capitalization changes made for the `case_sensitive` prelude flags, object names are no longer printed in upper case to the error log. However, this was causing some test cases to fail. I tried forcing the logs to print these object names in upper caps if `case_sensitive` is disabled, but there are simply too many points in the code that need to change for this to be effective (and I might miss a few). I also don't want to change the equality assertion that used in the tests to be case insensitive, as it might break test cases in the future that do use `case_insensitive`. So the new behaviour is accepted, and I've adjusted the tests to match the new capitalization. Might give issues with merging, though.