#31 Reverts a change from `503c83e - Added GOTO commands.` to fix bug
A small renaming issue caused a level to fail loading and be skipped if there was a message in-between but no newlines. This what was causing the Danish Pastry test to fail. I believe there was a reason it should be a newline and not "message", but I'm not sure if I remembered this in the right context though.

For the first time in PS+'s lifetime, all tests are passing!