#104 Fixes that AGAIN simulation twiddle logs could be visible
A side effect is that the twiddle debug now requires VERBOSE_LOGGING to be on. Apparantly verbose_logging is always disabled during again turn simulation. Which makes sense, but makes handling this case a bit weird, I guess. I've tried a more complex solution first, but I think this is simpler, and maybe the flag should be deprecated and twiddles should always be shown. Changed some warning messages for clarification